+++ src/components/AddButton/index.js
@@ -19,7 +19,7 @@ export default function AddButton() {
       aria-label="add"
       // onClick={() => props.addNewTrip("CreateTrip")}
       component={Link}
-      to="/newActivityForm"
+      to="/newTripForm"
     >
       <AddIcon />
     </Fab>
