+++ src/components/Expenses/index.js
@@ -1,35 +1,93 @@
-// import React from 'react';
-// import Header from './Header';
-// import Expenses from './Expenses';
+// import React, { useState } from "react";
+// import { makeStyles } from "@material-ui/core/styles";
+// import {
+//   TextField,
+//   Button,
+//   FormControl,
+//   InputLabel,
+//   InputAdornment,
+//   OutlinedInput,
+// } from "@material-ui/core/";
+// import { render } from "@testing-library/react";
 
-// import Form from './components/Form'
+// const useStyles = makeStyles((theme) => ({
+//   root: {
+//     display: "flex",
+//     flexWrap: "wrap",
+//   },
+//   margin: {
+//     margin: theme.spacing(1),
+//   },
+//   withoutLabel: {
+//     marginTop: theme.spacing(3),
+//   },
+//   textField: {
+//     width: "25ch",
+//   },
+// }));
 
 // function Expenses() {
-//   const [expenses, setExpenses] = useState(ALL_EXPENSES)
+//   const classes = useStyles();
+//   const [totalExpenses, setTotalExpenses] = useState();
+//   const [inputValue, setInputValue] = useState(null);
+//   const [inputActivity, setInputActivity] = useState(null);
+//   const [expenses, setExpenses] = useState([]);
+
+//   const handleClick = () => {
+//     // click should add amount to total amount
+//     const newExpense = {
+//       value: inputValue,
+//       activity: inputActivity,
+//     };
+
+//     const data = expenses;
+//     data.push(newExpense);
+
+//     setExpenses([...data]);
+
+//     let total = 0;
+//     data.forEach((datum) => {
+//       total += datum.value;
+//     });
+
+//     setTotalExpenses(total);
+//   };
+//   const renderExpensesRow = () => {
+//     return expenses.map((expense) => {
+//       return (
+//         <div>
+//           <div>Value: {expense.value}</div>
+//           <div>activity: {expense.activity}</div>
+//         </div>
+//       );
+//     });
+//   };
 
 //   return (
-//     <Container>
-//       <Jumbotron fluid>
-//         <h3 className='display-6 text-center'>
-//           Expense Tracker React App
-//           <img src={Logo} style={{ width: 50, height: 50 }} alt='react-logo' />
-//         </h3>
-//         <div className='text-center'>
-//           <p>
-//             Total Expense:{' '}
-//             <span className='text-success'>
-//               ${' '}
-//               {expenses.reduce((accumulator, currentValue) => {
-//                 return (accumulator += parseInt(currentValue.amount))
-//               }, 0)}
-//             </span>
-//           </p>
-//         </div>
-//         {*/ ADD THE BELOW LINE/*}
-//         <Form />
-//       </Jumbotron>
-// //     </Container>
-//   )
+//     <>
+//       <Box>Total: {totalExpenses}</Box>
+//       <TextField
+//         label="Name of Activity"
+//         id="filled-size-small"
+//         value={inputActivity ? inputActivity : ""}
+//         onChange={(event) => setInputActivity(event.target.value)}
+//       />
+//       <FormControl className={classes.margin} variant="outlined">
+//         <InputLabel htmlFor="outlined-adornment-amount">Amount</InputLabel>
+//         <OutlinedInput
+//           id="outlined-adornment-amount"
+//           onChange={(event) => setInputValue(parseInt(event.target.value))}
+//           value={inputValue ? inputValue : 0}
+//           startAdornment={<InputAdornment position="start">$</InputAdornment>}
+//           labelWidth={60}
+//         />
+//       </FormControl>
+
+//       <Button onClick={handleClick} />
+
+//       <div>{renderExpensesRow()}</div>
+//     </>
+//   );
 // }
 
-// export default App
+// export default Expenses;
