+++ src/pages/Expenses/index.js
@@ -1,72 +1,95 @@
-import React from "react";
-import Container from "@material-ui/core/Container";
-import Box from "@material-ui/core/box";
-// import Paper from "@material-ui/core/Paper";
-import Fab from "@material-ui/core/Fab";
-import AddIcon from "@material-ui/icons/Add";
-import Card from "@material-ui/core/Card";
-// import TripHeader from "../../components/TripHeader";
-import { Link, useParams } from "react-router-dom";
+import React, { useState } from "react";
+import { makeStyles } from "@material-ui/core/styles";
+import {
+  TextField,
+  Box,
+  Button,
+  FormControl,
+  InputLabel,
+  InputAdornment,
+  OutlinedInput,
+} from "@material-ui/core/";
+import { render } from "@testing-library/react";
 
-const containerStyle = {
-  backgroundColor: "white",
-  height: '100vh',
-  borderTopRightRadius: 0,
-  borderTopLeftRadius: 0,
-  borderBottomRightRadius: 8,
-  borderBottomLeftRadius: 8,
-  border: 0,
-  color: '#333333',
-  padding: 0,
-};
+const useStyles = makeStyles((theme) => ({
+  root: {
+    display: "flex",
+    flexWrap: "wrap",
+  },
+  margin: {
+    margin: theme.spacing(1),
+  },
+  withoutLabel: {
+    marginTop: theme.spacing(3),
+  },
+  textField: {
+    width: "25ch",
+  },
+}));
 
-const btnStyle = {
-  position: 'fixed',
-  bottom: '15px',
-  right: '15px',
-}
+function Expenses() {
+  const classes = useStyles();
+  const [totalExpenses, setTotalExpenses] = useState();
+  const [inputValue, setInputValue] = useState(null);
+  const [inputActivity, setInputActivity] = useState(null);
+  const [expenses, setExpenses] = useState([]);
+  
 
-export default function Expenses() {
-  let { id } = useParams();
+  const handleClick = () => {
+    // click should add amount to total amount
+    const newExpense = {
+      value: inputValue,
+      activity: inputActivity,
+    };
+
+    const data = expenses;
+    data.push(newExpense);
+
+    setExpenses([...data]);
+
+    let total = 0;
+    data.forEach((datum) => {
+      total += datum.value;
+    });
+
+    setTotalExpenses(total);
+  };
+  const renderExpensesRow = () => {
+    return expenses.map((expense) => {
+      return (
+        <div>
+          <div>Value: {expense.value}</div>
+          <div>activity: {expense.activity}</div>
+        </div>
+      );
+    });
+  };
 
   return (
-    <div style={{paddingLeft:15, paddingRight:15}}>
-      <Container maxWidth="md" style={containerStyle}>
-        <typography>
-          <h2 style={{margin:0, padding: '15px'}}> Explore Header</h2>
-        </typography>
-          {/* <TripHeader /> */}
-          <Box p={2}>
-            <Card>
-              <h3>Expenses content goes here</h3>
-            </Card>
-          </Box>
+    <>
+      <Box>Total: {totalExpenses}</Box>
+      <TextField
+        label="Name of Activity"
+        id="filled-size-small"
+        value={inputActivity ? inputActivity : ""}
+        onChange={(event) => setInputActivity(event.target.value)}
+      />
+      <FormControl className={classes.margin} variant="outlined">
+        <InputLabel htmlFor="outlined-adornment-amount">Amount</InputLabel>
+        <OutlinedInput
+          id="outlined-adornment-amount"
+          onChange={(event) => setInputValue(parseInt(event.target.value))}
+          value={inputValue ? inputValue : 0}
+          startAdornment={<InputAdornment position="start">$</InputAdornment>}
+          labelWidth={60}
+        />
+      </FormControl>
 
-          <Box p={2}>
-            <Card>
-              <h3>More expenses content goes here</h3>
-            </Card>
-          </Box>
+      <Button onClick={handleClick}>Submit</Button>
 
-          <div>
-            <Box
-              display="flex"
-              style={{ justifyContent: "flex-end", alignItems: "flex-end" }}
-              m={2}
-            >
-              <Fab
-                color="primary"
-                aria-label="add"
-                // onClick={() => props.addNewTrip("CreateTrip")}
-                component={Link}
-                to={`/Trip/${id}/Dashboard/newexpenseform`}
-              >
-                <AddIcon />
-              </Fab>
-            </Box>
-          </div>
-          
-      </Container>
-      </div>
+      <div>{renderExpensesRow()}</div>
+    </>
   );
 }
+
+export default Expenses;
